Giant ubercommit.
Includes a bunch of features, but I didn't commit regularly
so they're all mashed together. :(

the most important of them are changes to the API access to use
reqtify instead of a naked http client and GET requests.

also important is the change to allow fetching the response body reader
from an api call.