Tweak BuildAnswerInlineQueryReq to be greatly more space efficient
We were running up against telegram bot api size limitations, namely
that the body of a request is not allowed to exceed 64KB. This is
normally fine for everything but it's just slightly too low to serve
a full page of inline query results with big fluffy captions.

I am still unable to serve all 50, but with this optimization, I can
at least do 40 at a time, which is a long shot better than 15.